move join logic to separate module; name functions
