+++ lib/search.js
@@ -1,5 +1,6 @@
 const R = require('ramda')
 const Query = require('./query.js')
+const Join = require('./search/join.js')
 
 const PAGINATION_PARAMS = [
   'page',
@@ -12,8 +13,6 @@ const DEFAULT_PAGINATION_PARAMS = {
 }
 
 const PAGINATION_LIMIT_CLAUSE = ' LIMIT :offset, :limit'
-
-const REGEX_JOIN = /((?:LEFT|RIGHT|INNER|OUTER|CROSS)? ?JOIN +`?(.+?)`? *(?: +AS +`?(.+?)`?)? +ON .+?) ?(?=LEFT|RIGHT|INNER|OUTER|CROSS|JOIN|\n|$)/i
 
 const CLAUSES = [
   'select',
@@ -108,7 +107,7 @@ const getClauses = (map, params) => {
   }
 
   return R.compose(
-    _sanitizeJoins,
+    Join.sanitizeJoins,
     R.reduce(_mergeClauses, {}),
     R.values,
     R.mapObjIndexed(_resolveClause(params)),
@@ -126,59 +125,7 @@ const getCountSql = (sql) => {
 }
 
 
-// _getJoins:: String -> Array
-const _getJoins = (input) => {
 
-  let matches, output = []
-
-  let index = 0
-
-  matches = REGEX_JOIN.exec(input)
-
-  while(matches) {
-    var [match, table, alias] = matches
-
-    output.push({
-      alias: R.defaultTo(table, alias),
-      match: match,
-      index: matches.index + index
-    })
-
-    input = R.slice(matches.index+match.length, Infinity, matches.input)
-    index += matches.index+match.length
-
-    matches = REGEX_JOIN.exec(input)
-  }
-
-  return output
-
-}
-
-
-// _sanitizeJoins:: String -> Array -> String
-const _sanitizeJoins = (clauses) => {
-
-  let join_clause = clauses.join
-
-  let joins = _getJoins(join_clause)
-
-  if(R.isNil(join_clause))
-    return ''
-
-  return R.compose(
-    R.merge(clauses),
-    R.objOf('join'),
-    R.reduce((c,n) => {
-      return R.join('', R.remove(n.index, n.match.length, c))
-    }, join_clause),
-    R.sortBy(R.compose(R.negate, R.prop('index'))),
-    R.flatten,
-    R.map(R.tail),
-    R.values,
-    R.groupBy(R.prop('alias'))
-  )(joins)
-
-}
 
 module.exports = {
   parseParams: parseParams,
